Move down form status state
